+++ src/core/index.ts
@@ -1,7 +1,6 @@
 export * from "./types";
-export * from "./transform";
+export { parseTab, TAB_RENDERER_PARSER_VERSION, TAB_RENDERER_AST_VERSION } from "./parseTab";
 export * from "./preparedTypes";
-export { prepareSong } from "./prepareSong";
 export { chordToText } from "./transposer/chordToText";
 export { generateBarList } from "./renderer/generateBarList";
 export type { SectionText, SectionBarList, BarsListItem } from "./parser/types";